requirements.txt
numpy>=1.20.0
opencv-contrib-python>=4.5.5.64
opencv-python>=4.5.5.64
openpyxl>=3.0.9
pandas>=1.3.5
pillow>=9.0.0
pyttsx3>=2.90
tqdm>=4.64.0